test-2 refactor all tests into quest flow and add test change status tests
