+++ all_tests.sh
@@ -1,0 +1,3 @@
+#!/usr/bin/env bash
+
+npx playwright test tests/all_tests
