--- client_tests.sh
@@ -1,3 +1,0 @@
-#!/usr/bin/env bash
-
-npx playwright test tests/all_tests/client
